cli: remove circular import
